% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_library.R
\name{create_library}
\alias{create_library}
\title{create_library}
\usage{
create_library(url, repositoryName = "Foundation", serverName = NULL,
  libraryName, engine, displayName, path, location, isPreassigned,
  asDataFrame = FALSE)
}
\arguments{
\item{url}{URL of the server with installed SAS9API.}

\item{repositoryName}{Repository name.}

\item{serverName}{Server name.}

\item{libraryName}{Library name.}

\item{engine}{Library engine.}

\item{displayName}{Display name.}

\item{path}{Library data path.}

\item{location}{Folder to place metadata object in.}

\item{isPreassigned}{logical. TRUE if created library should be preassigned}

\item{asDataFrame}{logical. Determines the content of the response returned by the function. If FALSE, the function will return full JSON response. If TRUE, the function will return only payload part of the response transformed into a dataframe.}
}
\description{
Creates library at a given server with the given library name and parameters.
}
\examples{
create_library(url, serverName, libraryName = "new_lib_1",
engine = "V9", displayName = "New Library 1", path = "/pub/portal/data",
location = "/User Folders/User1", isPreassigned = TRUE)

}
