% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_library.R
\name{delete_library}
\alias{delete_library}
\title{delete_library}
\usage{
delete_library(url = NULL, repositoryName = "Foundation", serverName,
  libraryName, asDataFrame = FALSE)
}
\arguments{
\item{url}{URL of the server with installed SAS9API.}

\item{repositoryName}{Repository name.}

\item{serverName}{Server name.}

\item{libraryName}{Library name.}

\item{asDataFrame}{logical. Determines the content of the response returned by the function. If FALSE, the function will return full JSON response. If TRUE, the function will return only payload part of the response transformed into a dataframe.}
}
\description{
Removes all libraries with matching library name.
}
