% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_group_info.R
\name{get_group_info}
\alias{get_group_info}
\title{get_group_info}
\usage{
get_group_info(url, repositoryName = "Foundation", groupName,
  asDataFrame = FALSE)
}
\arguments{
\item{url}{URL of the server with installed SAS9API.}

\item{repositoryName}{Repository name.}

\item{groupName}{Group name.}

\item{asDataFrame}{logical. Determines the content of the response returned by the function. If FALSE, the function will return full JSON response. If TRUE, the function will return only payload part of the response transformed into a dataframe.}
}
\description{
Gets the group information and its associated groups and users by a group name.
}
