% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_library_info.R
\name{get_library_info}
\alias{get_library_info}
\title{get_library_info}
\usage{
get_library_info(url, repositoryName = "Foundation", serverName = NULL,
  serverUrl = NULL, serverPort = NULL, libraryName,
  asDataFrame = FALSE)
}
\arguments{
\item{url}{URL of the server with installed SAS9API.}

\item{repositoryName}{Repository name.}

\item{serverName}{Workspace server name.}

\item{serverUrl}{Workspace server URL.}

\item{serverPort}{Workspace server port.}

\item{libraryName}{Library name.}

\item{asDataFrame}{logical. Determines the content of the response returned by the function. If FALSE, the function will return full JSON response. If TRUE, the function will return only payload part of the response transformed into a dataframe.}
}
\description{
Gets the library information for the workspace server.
Workspace server can be connected either by server name or by server url with server port.
If none of these connections are provided, default workspace server from API configuration file will be used.
}
