% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata_server_config.R
\name{get_metadata_server_config}
\alias{get_metadata_server_config}
\title{get_metadata_server_config}
\usage{
get_metadata_server_config(url, asDataFrame = FALSE)
}
\arguments{
\item{url}{URL of the server with installed SAS9API.}

\item{asDataFrame}{logical. Determines the content of the response returned by the function. If FALSE, the function will return full JSON response. If TRUE, the function will return only payload part of the response transformed into a dataframe.}
}
\description{
Gets the current metadata server configuration.
}
