% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_role_info.R
\name{get_role_info}
\alias{get_role_info}
\title{get_role_info}
\usage{
get_role_info(url, repositoryName = "Foundation", roleName,
  asDataFrame = FALSE)
}
\arguments{
\item{url}{URL of the server with installed SAS9API.}

\item{repositoryName}{Repository name.}

\item{roleName}{Role name.}

\item{asDataFrame}{logical. Determines the content of the response returned by the function. If FALSE, the function will return full JSON response. If TRUE, the function will return only payload part of the response transformed into a dataframe.}
}
\description{
Gets the role information and its associated groups and users by a role name.
}
