% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_role_list.R
\name{get_role_list}
\alias{get_role_list}
\title{get_role_list}
\usage{
get_role_list(url, repositoryName = "Foundation", asDataFrame = FALSE)
}
\arguments{
\item{url}{URL of the server with installed SAS9API.}

\item{repositoryName}{Repository name.}

\item{asDataFrame}{logical. Determines the content of the response returned by the function. If FALSE, the function will return full JSON response. If TRUE, the function will return only payload part of the response transformed into a dataframe.}
}
\description{
Gets the list of roles and their associated groups and users.
}
