% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stp_server_config.R
\name{get_stp_server_config}
\alias{get_stp_server_config}
\title{get_stp_server_config}
\usage{
get_stp_server_config(url, repositoryName = "Foundation",
  serverName = NULL, asDataFrame = FALSE)
}
\arguments{
\item{url}{URL of the server with installed SAS9API.}

\item{repositoryName}{Repository name.}

\item{serverName}{Server name.}

\item{asDataFrame}{logical. Determines the content of the response returned by the function. If FALSE, the function will return full JSON response. If TRUE, the function will return only payload part of the response transformed into a dataframe.}
}
\description{
Gets the Stored Process Server information and its connections from the metadata server by a server name.
}
