% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user_info.R
\name{get_user_info}
\alias{get_user_info}
\title{get_user_info}
\usage{
get_user_info(url, repositoryName = "Foundation", userName,
  asDataFrame = FALSE)
}
\arguments{
\item{url}{URL of the server with installed SAS9API.}

\item{repositoryName}{Repository name.}

\item{userName}{User name.}

\item{asDataFrame}{logical. Determines the content of the response returned by the function. If FALSE, the function will return full JSON response. If TRUE, the function will return only payload part of the response transformed into a dataframe.}
}
\description{
Gets the server user information and its identities by user name.
}
