% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_workspace_server_list.R
\name{get_workspace_server_list}
\alias{get_workspace_server_list}
\title{get_workspace_server_list}
\usage{
get_workspace_server_list(url, repositoryName = "Foundation",
  asDataFrame = FALSE)
}
\arguments{
\item{url}{URL of the server with installed SAS9API.}

\item{repositoryName}{Repository name.}

\item{asDataFrame}{logical. Determines the content of the response returned by the function. If FALSE, the function will return full JSON response. If TRUE, the function will return only payload part of the response transformed into a dataframe.}
}
\description{
Gets a list of available workspace servers and their connections from the metadata server.
}
