% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_data.R
\name{insert_data}
\alias{insert_data}
\title{insert_data}
\usage{
insert_data(url = NULL, repositoryName = "Foundation",
  serverName = NULL, serverUrl = NULL, serverPort = NULL,
  libraryName, datasetName, byKey = NULL, data, asDataFrame = FALSE)
}
\arguments{
\item{url}{URL of the server with installed SAS9API.}

\item{repositoryName}{Repository name.}

\item{serverName}{Workspace server name. Either server name or server url with server port should be provided.}

\item{serverUrl}{Workspace server URL.}

\item{serverPort}{Workspace server port.}

\item{libraryName}{Library name.}

\item{datasetName}{Dataset name.}

\item{byKey}{Dataset key for record matching.}

\item{data}{Data in JSON format.}

\item{asDataFrame}{logical. Determines the content of the response returned by the function. If FALSE, the function will return full JSON response. If TRUE, the function will return only payload part of the response transformed into a dataframe.}
}
\description{
Inserts data into the dataset or replaces data by key.
}
\examples{
insert_data(url, serverUrl, serverPort, libraryName, datasetName,
data = "[{'Customer_Country': 'AU'}]")

}
