% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_object.R
\name{move_object}
\alias{move_object}
\title{move_object}
\usage{
move_object(url, repositoryName = "Foundation", sourceLocation,
  sourceName, publicType, destinationLocation, asDataFrame = FALSE)
}
\arguments{
\item{url}{URL of the server with installed SAS9API.}

\item{repositoryName}{Repository name.}

\item{sourceLocation}{Object location.}

\item{sourceName}{Object name.}

\item{publicType}{Object public type.}

\item{destinationLocation}{Destination location.}

\item{asDataFrame}{logical. Determines the content of the response returned by the function. If FALSE, the function will return full JSON response. If TRUE, the function will return only payload part of the response transformed into a dataframe.}
}
\description{
Moves object.
}
\examples{
move_object(url, sourceLocation = "User Folders/user1",
sourceName = "Library1", publicType = "Library", destinationLocation = "User Folders/user2")

}
