% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_all_data.R
\name{replace_all_data}
\alias{replace_all_data}
\title{replace_all_data}
\usage{
replace_all_data(url, repositoryName = "Foundation", serverName = NULL,
  serverUrl = NULL, serverPort = NULL, libraryName, datasetName, data,
  asDataFrame = FALSE)
}
\arguments{
\item{url}{URL of the server with installed SAS9API.}

\item{repositoryName}{Repository name.}

\item{serverName}{Workspace server name.}

\item{serverUrl}{Workspace server URL.}

\item{serverPort}{Workspace server port.}

\item{libraryName}{Library name.}

\item{datasetName}{Dataset name.}

\item{data}{Data in JSON format.}

\item{asDataFrame}{logical. Determines the content of the response returned by the function. If FALSE, the function will return full JSON response. If TRUE, the function will return only payload part of the response transformed into a dataframe.}
}
\description{
Replaces all data in the dataset with input data.
Workspace server can be connected either by server name or by server url with server port.
If none of these connections are provided, default workspace server from API configuration file will be used.
}
\examples{
replace_all_data(url, serverUrl, serverPort, libraryName, datasetName,
data = '[{"Name": "Andrew","Sex": "M","Age": 55,"Height": 184,"Weight": 91}]')

}
