% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_data.R
\name{retrieve_data}
\alias{retrieve_data}
\title{retrieve_data}
\usage{
retrieve_data(url, repositoryName = "Foundation", serverName = NULL,
  serverUrl = NULL, serverPort = NULL, libraryName, datasetName,
  limit = 100, offset = 0, filter = NULL, asDataFrame = FALSE)
}
\arguments{
\item{url}{URL of the server with installed SAS9API.}

\item{repositoryName}{Repository name.}

\item{serverName}{Workspace server name.}

\item{serverUrl}{Workspace server URL.}

\item{serverPort}{Workspace server port.}

\item{libraryName}{Library name.}

\item{datasetName}{Dataset name.}

\item{limit}{Number of records to retrieve (default is 100, maximum value is 10000).}

\item{offset}{Dataset record offset (default is 0).}

\item{filter}{Dataset filter in JSON format.}

\item{asDataFrame}{logical. Determines the content of the response returned by the function. If FALSE, the function will return full JSON response. If TRUE, the function will return only payload part of the response transformed into a dataframe.}
}
\description{
Retrieves data from the dataset by a dataset name and a library name.
Workspace server can be connected either by server name or by server url with server port.
If none of these connections are provided, default workspace server from API configuration file will be used.
}
