% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_response.R
\name{return_response}
\alias{return_response}
\title{return_response}
\usage{
return_response(response, asDataFrame)
}
\arguments{
\item{response}{Response from API request.}

\item{asDataFrame}{logical. Determines the content of the response returned by the function. If FALSE, the function will return full JSON response. If TRUE, the function will return only payload part of the response transformed into a dataframe.}
}
\description{
Base function for handling API response.
}
