% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset_info.R
\name{get_dataset_info}
\alias{get_dataset_info}
\title{get_dataset_info}
\usage{
get_dataset_info(url, serverName = NULL, serverUrl = NULL,
  serverPort = NULL, libraryName, datasetName, asDataFrame = FALSE)
}
\arguments{
\item{url}{URL of the server with installed SAS9API.}

\item{serverName}{Workspace server name.}

\item{serverUrl}{Workspace server URL.}

\item{serverPort}{Workspace server port.}

\item{libraryName}{Library name.}

\item{datasetName}{Dataset name.}

\item{asDataFrame}{logical. Determines the content of the response returned by the function. If FALSE, the function will return full JSON response. If TRUE, the function will return only payload part of the response transformed into a dataframe.}
}
\description{
Gets the dataset information by defined library name and dataset name.
Workspace server can be connected either by server name or by server url with server port.
If none of these connections are provided, default workspace server from API configuration file will be used.
}
