% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_library_list.R
\name{get_library_list}
\alias{get_library_list}
\title{get_library_list}
\usage{
get_library_list(url, repositoryName = "Foundation", serverName = NULL,
  serverUrl = NULL, serverPort = NULL, asDataFrame = FALSE)
}
\arguments{
\item{url}{URL of the server with installed SAS9API.}

\item{repositoryName}{Repository name.}

\item{serverName}{Workspace server name. Either server name or server url with server port should be provided.}

\item{serverUrl}{Workspace server URL.}

\item{serverPort}{Workspace server port.}

\item{asDataFrame}{logical. Determines the content of the response returned by the function. If FALSE, the function will return full JSON response. If TRUE, the function will return only payload part of the response transformed into a dataframe.}
}
\description{
Gets the list of libraries for the workspace server.
}
